+++ controller/src/controller/constants.py
@@ -236,7 +236,7 @@ STIM_SHORT_CIRCUIT_THRESHOLD_OHMS = 10
 
 # Stim Subprotocols
 VALID_STIMULATION_TYPES = frozenset(["C", "V"])
-VALID_SUBPROTOCOL_TYPES = frozenset(["delay", "monophasic", "biphasic"])
+VALID_SUBPROTOCOL_TYPES = frozenset(["delay", "monophasic", "biphasic", "loop"])
 
 # does not include subprotocol idx
 STIM_PULSE_BYTES_LEN = 29